requirements.txt
pytest
mysql-connector-python
bcrypt
python-dotenv
flake8